#5914. As a followup to previous commit, don't round text sizes to integer without need.
idDeviceContext has a bunch of methods like: CharWidth, TextWidth, TextHeight, MaxCharWidth, MaxCharHeight.
For some reason they were roudning their result down.
That might be OK for getting size of single text, but that's not enough for quering individual characters.

This is more like a cleaning change to make it harder to fuckup text sizing by summing CharWidth or estimating max number of character by dividing on MaxCharWidth, etc.
I hope it won't break anything.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10233 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ ui/DeviceContext.h
@@ -48,11 +48,11 @@ public:
 	void				DrawMaterialRotated(float x, float y, float w, float h, const idMaterial *mat, const idVec4 &color, float scalex = 1.0, float scaley = 1.0, float angle = 0.0f);
 	void				DrawStretchPicRotated(float x, float y, float w, float h, float s0, float t0, float s1, float t1, const idMaterial *mat, float angle = 0.0f);
 
-	int					CharWidth( const char c, float scale );
-	int					TextWidth(const char *text, float scale, int limit);
-	int					TextHeight(const char *text, float scale, int limit);
-	int					MaxCharHeight(float scale);
-	int					MaxCharWidth(float scale);
+	float				CharWidth( const char c, float scale );
+	float				TextWidth(const char *text, float scale, int limit);
+	float				TextHeight(const char *text, float scale, int limit);
+	float				MaxCharHeight(float scale);
+	float				MaxCharWidth(float scale);
 
 	int					FindFont( const char *name );
 	void				SetupFonts();
